Move initMutationObserver
