+++ packages/rum/src/domain/record/observers/mutationObserver.ts
@@ -49,6 +49,14 @@ export type RumMutationRecord =
   | RumAttributesMutationRecord
   | RumChildListMutationRecord
 
+export function initMutationObserver(
+  cb: MutationCallBack,
+  configuration: RumConfiguration,
+  shadowRootsController: ShadowRootsController
+) {
+  return startMutationObserver(cb, configuration, shadowRootsController, document)
+}
+
 /**
  * Buffers and aggregate mutations generated by a MutationObserver into MutationPayload
  */
